+++ example/lib/pages/signin_page.dart
@@ -10,13 +10,13 @@ import 'package:getxfire/getxfire.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_signin_button/flutter_signin_button.dart';
-// import 'package:google_sign_in/google_sign_in.dart';
-// import 'package:flutter/foundation.dart' show kIsWeb;
+import 'package:flutter/foundation.dart' show kIsWeb;
 
 part 'auth/update_user.dart';
 part 'auth/user_infocard.dart';
 part 'auth/email_password_form.dart';
 part 'auth/anonymously.dart';
+part 'auth/other_providers_signin.dart';
 
 /// Entrypoint example for various sign-in flows with Firebase.
 class SignInPage extends StatefulWidget {
@@ -41,24 +41,41 @@ class _SignInPageState extends State<SignInPage> {
         title: Text(widget.title),
         actions: <Widget>[
           Builder(builder: (BuildContext context) {
-            return FlatButton(
-              textColor: Theme.of(context).buttonColor,
-              onPressed: () async {
-                final User user = GetxFire.currentUser;
-                if (user == null) {
-                  ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-                    content: Text('No one has signed in.'),
-                  ));
-                  return;
-                }
-                await GetxFire.signOut(isSocialLogout: false);
+            return Padding(
+              padding: const EdgeInsets.only(right: 8.0),
+              child: FlatButton(
+                textColor: Theme.of(context).buttonColor,
+                onPressed: () async {
+                  GetxFire.openDialog.confirm(
+                    content: "Are you sure to sign out?",
+                    // lottiePath: "assets/lottie/coffee-favorite.json",
+                    lottiePath: GetxFire.lottiePath.THINKING,
+                    onYesClicked: () async {
+                      final User user = GetxFire.currentUser;
+                      if (user == null) {
+                        ScaffoldMessenger.of(context)
+                            .showSnackBar(const SnackBar(
+                          content: Text('No one has signed in.'),
+                        ));
+                        return;
+                      }
+                      await GetxFire.signOut();
 
-                final String uid = user.uid;
-                ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  content: Text('$uid has successfully signed out.'),
-                ));
-              },
-              child: const Text('Sign out'),
+                      final String uid = user.uid;
+                      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                        content: Text('$uid has successfully signed out.'),
+                      ));
+                    },
+                  );
+                },
+                child: Text(
+                  'Sign Out',
+                  style: TextStyle(
+                    fontSize: 17,
+                    color: Colors.red[400],
+                  ),
+                ),
+              ),
             );
           })
         ],
@@ -72,7 +89,7 @@ class _SignInPageState extends State<SignInPage> {
             // EmailLinkSignInSection(),
             AnonymouslySignInSection(),
             // PhoneSignInSection(Scaffold.of(context)),
-            // OtherProvidersSignInSection(),
+            OtherProvidersSignInSection(),
           ],
         );
       }),