+++ lib/src/lottie_path.dart
@@ -1,0 +1,22 @@
+part of '../getxfire.dart';
+
+class LottiePath {
+  // ignore: non_constant_identifier_names
+  final THINKING = "assets/lottie/thinking.json";
+  // ignore: non_constant_identifier_names
+  final COMING_SOON = "assets/lottie/coming-soon.json";
+  // ignore: non_constant_identifier_names
+  final SEARCH_FILES = "assets/lottie/search-files.json";
+  // ignore: non_constant_identifier_names
+  final SAD_HEART = "assets/lottie/sad-heart.json";
+  // ignore: non_constant_identifier_names
+  final NO_ACTIVITY = "assets/lottie/no-activity.json";
+  // ignore: non_constant_identifier_names
+  final MULTI_TASKING = "assets/lottie/multi-tasking.json";
+  // ignore: non_constant_identifier_names
+  final IMAGE_ICON = "assets/lottie/image-icon.json";
+  // ignore: non_constant_identifier_names
+  final EMPTY_BOX = "assets/lottie/empty-box.json";
+  // ignore: non_constant_identifier_names
+  final COFFEE_FAVORITE = "assets/lottie/coffee-favorite.json";
+}